feat!: Users and documentation are encrypted
BREAKING CHANGE: The Person entity has been replaced by User with JWT.
ALERT: Unexpected change prevents access to documentation in Swagger
TODO: Fix documentation

+++ src/main/java/com/upc/TuCine/shared/exception/ResourceNotFoundException.java
@@ -14,7 +14,7 @@ public class ResourceNotFoundException extends RuntimeException{
         super(message, cause);
     }
 
-    public ResourceNotFoundException(String resourceName, Long resourceId) {
+    public ResourceNotFoundException(String resourceName, Integer resourceId) {
         super(String.format("%s with id %d not found.", resourceName, resourceId));
     }
 }
